requirements.txt
torch>=1.9.0
torchvision>=0.10.0
albumentations>=1.0.0
Pillow>=8.0.0
numpy>=1.21.0
tqdm>=4.62.0
